chore(components): rename PostmanRunButton
